backend/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.32
packaging==24.1
pluggy==1.5.0
protobuf==3.20.3
PyMySQL==1.1.1
pytest==7.4.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.3.7